Split up index.js into 3 files.
